published test

+++ US_Aerial/Autoprep/GetAerialnewreworkPNG.py
@@ -1,6 +1,7 @@
 import os, arcpy, shutil
 import cx_Oracle
 import json
+from PIL import Image
 
 class Machine:
     machine_test = r"\\cabcvan1gis006"
@@ -127,6 +128,7 @@ if __name__ == '__main__':
     scratch = arcpy.env.scratchFolder#r'C:\Users\JLoucks\Documents\JL\test2'#arcpy.env.scratchFolder
     job_directory = r'\\192.168.136.164\v2_usaerial\JobData\test'
     arcpy.env.OverwriteOutput = True
+    Image.MAX_IMAGE_PIXELS = 10000000000
 
     orderInfo = Oracle('test').call_function('getorderinfo',OrderID)
     OrderNumText = str(orderInfo['ORDER_NUM'])
@@ -161,15 +163,20 @@ if __name__ == '__main__':
                     """Two copies are performed, 1 to convert the raster into a PNG for the application.
                     the other to only copy the image without spatial information to the job folder"""
                     #arcpy.CopyRaster_management(imageuploadpath,os.path.join(scratch,job_image_name),colormap_to_RGB='ColormapToRGB',pixel_type='8_BIT_UNSIGNED',format='PNG',transform='NONE')
-                    arcpy.env.compression = "JPEG 50"
-                    arcpy.CopyRaster_management(originalpath,os.path.join(scratch,job_image_name),colormap_to_RGB='ColormapToRGB',pixel_type='8_BIT_UNSIGNED',format='JPEG',transform='NONE')
-                    if os.path.exists(os.path.join(job_folder,'gc',job_image_name)):
-                        os.remove(os.path.join(job_folder,'gc',job_image_name))
-                    shutil.copy(os.path.join(scratch,job_image_name),os.path.join(job_folder,'gc',job_image_name))
+                    if imageuploadpath[-3:] in ['jpg','png','tif']:
+                        im = Image.open(imageuploadpath)
+                        subject_im = im.convert("RGB")
+                        subject_im.save(os.path.join(job_folder,'gc',job_image_name))
+                    else:
+                        arcpy.env.compression = "JPEG 50"
+                        arcpy.CopyRaster_management(originalpath,os.path.join(scratch,job_image_name),colormap_to_RGB='ColormapToRGB',pixel_type='8_BIT_UNSIGNED',format='JPEG',transform='NONE')
+                        if os.path.exists(os.path.join(job_folder,'gc',job_image_name)):
+                            os.remove(os.path.join(job_folder,'gc',job_image_name))
+                        shutil.copy(os.path.join(scratch,job_image_name),os.path.join(job_folder,'gc',job_image_name))
                     rename_call = str({"PROCEDURE":Oracle.erisapi_procedures['setimagename'],"ORDER_NUM":OrderNumText,"AUI_ID":auid,"IMAGE_NAME":str(new_image_name)})
                     rename_return = Oracle('test').call_erisapi(rename_call)
                     print json.loads(rework_return[1])
                 elif not os.path.exists(originalpath):
                     arcpy.AddError('cannot find image in inventory to convert, PLEASE CHECK PATH: '+originalpath)
     except Exception as e:
-        arcpy.AddError('Issue converting image: '+e.message)
+        arcpy.AddError('Issue converting image: '+e.message, imageuploadpath)
